Safari-609.1.11.1/LayoutTests/css-custom-properties-api/inherits-expected.txt

Specified in parent, inherits=true

test

Specified in parent, inherits=false

test

Specified in parent, not registered

test

Initial

test

Unset, inherits=true

test

Unset, inherits=false

test

A cycle between an inherits=true and inherits=false property

test

A cycle between an inherits=true and inherits=false property

test

A cycle between an inherits=true and inherits=false property with fallback

test

Inheritance should not create a cycle

test

A cycle between two unregistered properties

test

Revert, inherits=true

test

Revert, inherits=false

test

Revert, unregistered

test

Inherit, unregistered

test

Test that inherited properties do variable substitution before being inherited - registered

test

Test that inherited properties do variable substitution before being inherited - unregistered

test

No initial value in registered property should act like unregistered

test

(unregistered)

test

Inherit should be substituted for unregistered property

test


PASS Registration is successful 
PASS JS Attributes are valid for element 1 
PASS JS Attributes are valid for element 2 
PASS JS Attributes are valid for element 3 
PASS JS Attributes are valid for element 4 
PASS JS Attributes are valid for element 5 
PASS JS Attributes are valid for element 6 
PASS JS Attributes are valid for element 7 
PASS JS Attributes are valid for element 8 
PASS JS Attributes are valid for element 9 
PASS JS Attributes are valid for element 10 
PASS JS Attributes are valid for element 11 
PASS JS Attributes are valid for element 12 
PASS JS Attributes are valid for element 13 
PASS JS Attributes are valid for element 14 
PASS JS Attributes are valid for element 14-1 
PASS JS Attributes are valid for element 15 
PASS JS Attributes are valid for element 16 
PASS JS Attributes are valid for element 17 
PASS JS Attributes are valid for element 18 
PASS JS Attributes are valid for element 19 
